feat(264): hoist StoryboardPlaybackService + deliver remaining deferred tasks
Closes out the 16 tasks deferred from the Milestone B commit (f3e432d).

T-HOIST (T010-T015) — StoryboardPlaybackService lifted from
apps/vscode/src/services/storyboardPlayback.ts (983 lines) to
shared/components/src/storyboardPlayback/service.ts:
- New HostEventEmitter<T> / HostEvent / HostDisposable abstractions
  in events.ts that are structurally compatible with vscode.EventEmitter /
  vscode.Event / vscode.Disposable. The VS Code app's vscode.EventEmitter
  instances pass through unchanged.
- Port interfaces extracted into ports.ts.
- The three vscode-backed defaults (showErrorMessage, setContext,
  showInformationMessage) lifted to the instantiation site in extension.ts;
  the shared service has no vscode dependency.
- plotFromFeatures + featuresFromPlot moved to shared/components/src/
  storyboardPlayback/plotFromFeatures.ts; the apps/vscode/ copies are
  thin re-export shims.
- New @debrief/components/storyboardPlayback subpath export (esbuild +
  vitest aliases wired).
- Net-zero behaviour change: 840 pre-existing vscode tests (including
  49 storyboardPlayback tests) pass unchanged.

Other deferred tasks:
- T018 + T019 — MapView Storybook story "BriefingTileLayerProps" + a
  shared/components/e2e Playwright spec capturing it in light/dark/vscode
  themes (3 PNGs in evidence/screenshots/).
- T059 — boot.ts module extracted from App.tsx with 3 unit tests.
- T072 + T073 — ModeToggle vitest (6 cases) + TimeSlider vitest (7 cases).
- T074 + T075 — story-mode (?story=transport-bar|mode-toggle) in
  briefing-renderer's App; Playwright spec captures TransportBar idle +
  end-of-storyboard, ModeToggle minimal + present (4 PNGs in evidence).
- T079 — end-to-end Playwright spec: real export → real unzip into tmp
  → load index.html from file:// → advance through 4 Scenes → press
  Replay → assert 0 external requests.
- T086 — interaction GIF captured via Playwright recordVideo + ffmpeg
  conversion (~100 KB, 4 s at 12 fps; well under the 2 MB / 5 s budget).

Verification: 935 tests pass across the feature surface (840 vscode +
58 briefing-renderer + 31 MapView + 6 storybook = 935), 19 Playwright
specs (16 briefing-renderer + 3 shared/components Storybook) all green.

Shipped blog post updated with all new screenshots and the interaction
GIF; sample-briefing.zip regenerated; webview-e2e-summary + test-summary
refreshed.

Only T089 (/speckit.pr) remains — user-triggered.
