+++ apps/briefing-renderer/package.json
@@ -27,6 +27,7 @@
   },
   "devDependencies": {
     "@playwright/test": "^1.58.0",
+    "jszip": "^3.10.1",
     "@sparticuz/chromium": "^143.0.4",
     "@testing-library/dom": "^10.4.1",
     "@testing-library/react": "^14.0.0",
