+++ shared/components/package.json
@@ -17,6 +17,11 @@
       "import": "./dist/index.js",
       "require": "./dist/index.cjs"
     },
+    "./storyboardPlayback": {
+      "types": "./dist/storyboardPlayback/index.d.ts",
+      "import": "./dist/index.js",
+      "require": "./dist/index.cjs"
+    },
     "./MapView": {
       "types": "./dist/MapView/index.d.ts",
       "import": "./dist/MapView/index.js"
